Moved formframe to the borrowable record
Encountered a potential blocker however: currently the rust compiler
has a bug that prevents it from correctly infering lifetimes for objects
that contain multiple lifetimes.

The current workaround is by boxing the problem future, because this
hides it behind a &dyn Future. However this has obvious runtime costs,
both in extra allocations and in dynamic dispatch